Backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
google-generativeai==0.3.2
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.9
typing-extensions==4.9.0
pdfplumber>=0.10.2
python-docx>=1.0.0
pymongo>=4.6.0 